implement config validation for azure ad
