+++ accesshandler/pkg/providers/azure/ad/validate.go
@@ -4,7 +4,10 @@ import (
 	"context"
 	"encoding/json"
 
+	"github.com/common-fate/granted-approvals/accesshandler/pkg/diagnostics"
+	"github.com/common-fate/granted-approvals/accesshandler/pkg/providers"
 	"github.com/hashicorp/go-multierror"
+	"github.com/pkg/errors"
 )
 
 type ADErr struct {
@@ -64,4 +67,39 @@ func (p *Provider) Validate(ctx context.Context, subject string, args []byte) er
 	}
 
 	return result
+}
+func (p *Provider) TestConfig(ctx context.Context) error {
+	_, err := p.ListUsers(ctx)
+	if err != nil {
+		return errors.Wrap(err, "failed to list users while testing azure provider configuration")
+	}
+	_, err = p.ListGroups(ctx)
+	if err != nil {
+		return errors.Wrap(err, "failed to list groups while testing azure provider configuration")
+	}
+	return nil
+}
+func (p *Provider) ValidateConfig() map[string]providers.ConfigValidationStep {
+	return map[string]providers.ConfigValidationStep{
+		"list-users": {
+			Name: "List Azure AD users",
+			Run: func(ctx context.Context) diagnostics.Logs {
+				u, err := p.ListUsers(ctx)
+				if err != nil {
+					return diagnostics.Error(err)
+				}
+				return diagnostics.Info("Azure AD returned %d users", len(u))
+			},
+		},
+		"list-groups": {
+			Name: "List Azure AD groups",
+			Run: func(ctx context.Context) diagnostics.Logs {
+				g, err := p.ListGroups(ctx)
+				if err != nil {
+					return diagnostics.Error(err)
+				}
+				return diagnostics.Info("Azure AD returned %d groups", len(g))
+			},
+		},
+	}
 }
