+++ accesshandler/pkg/providers/okta/okta.go
@@ -3,13 +3,9 @@ package okta
 import (
 	"context"
 
-	"github.com/common-fate/granted-approvals/accesshandler/pkg/diagnostics"
-	"github.com/common-fate/granted-approvals/accesshandler/pkg/providers"
 	"github.com/common-fate/granted-approvals/pkg/gconfig"
 	"github.com/invopop/jsonschema"
 	"github.com/okta/okta-sdk-golang/v2/okta"
-	"github.com/okta/okta-sdk-golang/v2/okta/query"
-	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
 
@@ -38,43 +34,6 @@ func (o *Provider) Init(ctx context.Context) error {
 
 	o.client = client
 	return nil
-}
-
-func (p *Provider) TestConfig(ctx context.Context) error {
-	_, _, err := p.client.User.ListUsers(ctx, &query.Params{})
-	if err != nil {
-		return errors.Wrap(err, "failed to list users while testing okta provider configuration")
-	}
-	_, _, err = p.client.Group.ListGroups(ctx, &query.Params{})
-	if err != nil {
-		return errors.Wrap(err, "failed to list groups while testing okta provider configuration")
-	}
-	return nil
-}
-
-func (p *Provider) ValidateConfig() map[string]providers.ConfigValidationStep {
-	return map[string]providers.ConfigValidationStep{
-		"list-users": {
-			Name: "List Okta users",
-			Run: func(ctx context.Context) diagnostics.Logs {
-				u, _, err := p.client.User.ListUsers(ctx, &query.Params{})
-				if err != nil {
-					return diagnostics.Error(err)
-				}
-				return diagnostics.Info("Okta returned %d users (more may exist, pagination has been ignored)", len(u))
-			},
-		},
-		"list-groups": {
-			Name: "List Okta groups",
-			Run: func(ctx context.Context) diagnostics.Logs {
-				g, _, err := p.client.Group.ListGroups(ctx, &query.Params{})
-				if err != nil {
-					return diagnostics.Error(err)
-				}
-				return diagnostics.Info("Okta returned %d groups (more may exist, pagination has been ignored)", len(g))
-			},
-		},
-	}
 }
 
 // ArgSchema returns the schema for the Okta provider.
